pyproject.toml
[tool.poetry]
name = "google-drive-backup-cleaner"
version = "0.1.0"
description = ""
authors = ["ninecloud <gabrielarlo11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.128.0"
google-auth-oauthlib = "^1.2.0"
schedule = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
